+++ src/DrawingLink.UI/GH/GrasshopperCaller.cs
@@ -1,5 +1,4 @@
-﻿using DrawingLink.UI.Model;
-using GH_IO.Serialization;
+﻿using GH_IO.Serialization;
 using Grasshopper.Kernel;
 using Grasshopper.Kernel.Special;
 using Grasshopper.Kernel.Types;
