+++ src/DrawingLink.UI/GH/Models/InfoParam.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace DrawingLink.UI.GH.Models
+{
+    public class InfoParam : Param
+    {
+        public string Value { get; }
+
+        public InfoParam(string value, float top) : base(string.Empty, top)
+        {
+            Value = value ?? throw new ArgumentNullException(nameof(value));
+        }
+    }
+}
