+++ src/DrawingLink.UI/ParameterViewer.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using DrawingLink.UI.GH;
+using DrawingLink.UI.GH.Models;
 using System;
 using System.Diagnostics;
 using System.Linq;
